:truck: chore: moved misc components to shared module
