+++ app/src/main/java/com/shub39/grit/app/App.kt
@@ -46,12 +46,12 @@ import com.shub39.grit.app.AppSections.Companion.toIconRes
 import com.shub39.grit.app.AppSections.Companion.toStringRes
 import com.shub39.grit.billing.PaywallPage
 import com.shub39.grit.core.LocalWindowSizeClass
+import com.shub39.grit.core.components.ChangelogSheet
 import com.shub39.grit.core.domain.MainAppState
 import com.shub39.grit.core.domain.Sections
 import com.shub39.grit.core.habits.presentation.ui.HabitsGraph
 import com.shub39.grit.core.navigation.fadeTransitionMetadata
 import com.shub39.grit.core.navigation.verticalTransitionMetadata
-import com.shub39.grit.core.presentation.ChangelogSheet
 import com.shub39.grit.core.presentation.settings.ui.SettingsGraph
 import com.shub39.grit.core.tasks.presentation.ui.TasksPage
 import com.shub39.grit.viewmodels.HabitViewModel
