+++ app/src/main/java/com/shub39/grit/core/presentation/settings/ui/section/LookAndFeelPage.kt
@@ -67,7 +67,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.materialkolor.PaletteStyle
 import com.materialkolor.rememberDynamicColorScheme
-import com.shub39.grit.core.presentation.component.ColorPickerDialog
+import com.shub39.grit.core.components.ColorPickerDialog
 import com.shub39.grit.core.presentation.settings.SettingsAction
 import com.shub39.grit.core.presentation.settings.SettingsState
 import com.shub39.grit.core.shared_ui.detachedItemShape
