--- app/src/main/java/com/shub39/grit/core/presentation/util.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2026  Shubham Gorai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.shub39.grit.core.presentation
-
-private val brainrots =
-    listOf(
-        "Bombardino Crocodilo",
-        "Brr Brr Patapim",
-        "Lirili Larila",
-        "Trippi Troppi",
-        "Capucino Assassaino",
-        "Trulimero Trulichina",
-        "Tung Tung Tung Sahur",
-        "Chimpanzini Bananini",
-        "Giraffa Celeste",
-        "Tralalero Tralala",
-    )
-
-fun getRandomLine(): String = brainrots.random()
