+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/components/InitialLoading.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.presentation.component
+package com.shub39.grit.core.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
